Refactor figure generation scripts and enhance data validation
- Removed unused imports and functions in `generate_figures.py`.
- Updated figure generation logic to improve layout and aesthetics.
- Changed color maps in visualizations for better clarity.
- Enhanced logging messages for better debugging and information.
- Improved data validation in `preprocess_dataset.py` to ensure masks contain only valid values.
- Added checks for missing entropy columns in `run_experiment.py` to trigger precomputation if necessary.
- Updated aggregation functions in `aggregation.py` to use observed groupings for better performance.
- Enhanced correlation analysis in `statistics.py` to handle missing data more gracefully.
- Improved handling of edge cases in spatial autocorrelation calculations.

+++ src/pdi_pipeline/entropy.py
@@ -60,7 +60,7 @@ def shannon_entropy(
     image = np.asarray(image, dtype=np.float32)
 
     if image.ndim == 3:
-        gray = np.mean(image, axis=2)
+        gray = np.mean(image, axis=2, dtype=np.float32)
     elif image.ndim == 2:
         gray = image
     else:
